// Project/dff_library/ui_components/components/Toasts/utils/enum.ts
export enum ToastTypeValues {
  Success = 'Success',
  Error = 'Error',
  DefaultLeftIcon = 'DefaultLeftIcon',
  DefaultCloseTag = 'DefaultCloseTag',
  SuccessEphemeral = 'SuccessEphemeral',
  ErrorEphemeral = 'ErrorEphemeral',
  DefaultStaticLeftIcon = 'DefaultStatic',
  DefaultStaticClosingTag = 'DefaultStaticClosingTag',
}